Reorganize entire project
