+++ app/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 
-import Main from './components/new/Main';
+import Main from './components/Main';
 
 import './App.css';
