--- app/src/components/Popup.css
@@ -1,40 +1,0 @@
-.popup-box  {
-    color: #000000;
-    position: fixed;
-    background: #00000050;
-    width: 100%;
-    height: 100vh;
-    top: 0;
-    left: 0;
-  }
-   
-/* TODO: style the box and close icon consistently with rest of app */
-.box  {
-  position: relative;
-  width: 70%;
-  margin: 0 auto;
-  height: auto;
-  max-height: 70vh;
-  margin-top: calc(100vh - 85vh - 20px);
-  background: rgb(243, 238, 238);
-  border-radius: 4px;
-  padding: 20px;
-  border: 1px solid rgb(0, 0, 0);
-  overflow: auto;
-}
-  
-.close-icon  {
-  content: 'x';
-  cursor: pointer;
-  position: fixed;
-  right: calc(15% - 30px);
-  top: calc(100vh - 85vh - 33px);
-  background: #ededed;
-  width: 25px;
-  height: 25px;
-  border-radius: 50%;
-  line-height: 20px;
-  text-align: center;
-  border: 1px solid #999;
-  font-size: 20px;
-}
