--- app/src/components/SelectMode.css
@@ -1,10 +1,0 @@
-.SelectMode {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-    align-items: center;
-}
-
-.SelectModeHeader {
-    text-align: center;
-}
