--- app/src/components/YouTubePlayer.css
@@ -1,1 +1,0 @@
-.YouTubePlayer {}
