+++ app/src/components/content/ColorPalette.css
