+++ app/src/components/content/ColorPalette.tsx
