+++ app/src/components/content/SongPage.css
@@ -4,3 +4,11 @@
     grid-template-rows: minmax(0, 1fr) minmax(0, 1fr) 10%;
     height: 10in;
 }
+
+.lyrics {
+    display: flex;
+}
+
+.lyrics > * {
+    flex: 1;
+}
