+++ app/src/components/layout/sidebar/NavItem.tsx
