+++ app/src/components/layout/sidebar/Sidebar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Settings from './Settings';
 
-import { Settings as SettingsValues } from '../../util/settings';
+import { Settings as SettingsValues } from '../../../util/settings';
 
 import styles from './Sidebar.module.css';
 import SideNav from './SideNav';
