+++ app/src/components/lyrics/Lyrics.tsx
