+++ app/src/components/lyrics/LyricsLine.tsx
