+++ app/src/components/pieces/SmallCapsHeader.css
