+++ app/src/components/pieces/SmallCapsHeader.tsx
