+++ app/src/components/playback/Scrubber.tsx
