+++ app/src/util/config.ts
@@ -1,0 +1,5 @@
+export const config = {
+    TRANSLATE_KEY: '',
+};
+
+export default config;
