+++ app/src/util/translate.ts
